Refactored paging component into separate module
