+++ site/pages/meet-our-team/team-slice/paging/style.css
@@ -1,0 +1,19 @@
+@value badgerWhite, badgerBlack from "../../../../css/_colors.css";
+
+.paging {
+  display: flex;
+  justify-content: space-between;
+}
+
+.pagingButton {
+  composes: fontXS from "../../../../css/typography/_fonts.css";
+  composes: boldSansSerif from "../../../../css/typography/_fonts.css";
+  padding: 10px 20px;
+  background: badgerBlack;
+  color: badgerWhite;
+}
+
+.pagingButton:not([href]) {
+  background-color: badgerBlack;
+  opacity: 0.15;
+}
